+++ uliweb/contrib/i18n/middle_i18n.py
@@ -2,6 +2,8 @@ import re
 
 from uliweb.i18n import set_language, format_locale
 from uliweb import Middleware
+from logging import getLogger
+from uliweb.utils.common import request_url
 
 accept_language_re = re.compile(r'''
         ([A-Za-z]{1,8}(?:-[A-Za-z]{1,8})*|\*)   # "en", "en-au", "x-y-z", "*"
@@ -12,26 +14,40 @@ accept_language_re = re.compile(r'''
 def get_language_from_request(request, settings):
     #check query_string, and the key will be defined in settings.ini
     #now only support GET method
+    debug = '__debug__' in request.GET
+    log = getLogger(__name__)
     url_lang_key = settings.get_var('I18N/URL_LANG_KEY')
     if url_lang_key:
         lang = request.GET.get(url_lang_key)
         if lang:
+            if debug:
+                log.info('Detect from URL=%s, lang_key=%s, lang=%s' %
+                         (request_url(), url_lang_key, lang))
             return lang
 
     #check session
     if hasattr(request, 'session'):
         lang = request.session.get('uliweb_language')
         if lang:
+            if debug:
+                log.info('Detect from session=%s, lang=%s' %
+                         ('uliweb_language', lang))
             return lang
 
     #check cookie
     lang = request.cookies.get(settings.I18N.LANGUAGE_COOKIE_NAME)
     if lang:
+        if debug:
+            log.info('Detect from cookie=%s, lang=%s' %
+                     (settings.I18N.LANGUAGE_COOKIE_NAME, lang))
         return lang
 
     #check browser HTTP_ACCEPT_LANGUAGE head
     accept = request.environ.get('HTTP_ACCEPT_LANGUAGE', None)
     if not accept:
+        if debug:
+            log.info('Detect from settings=%s, lang=%s' %
+                     ('LANGUAGE_CODE', lang))
         return settings.I18N.get('LANGUAGE_CODE')
     languages = settings.I18N.get('SUPPORT_LANGUAGES', [])
     for accept_lang, unused in parse_accept_lang_header(accept):
@@ -43,6 +59,9 @@ def get_language_from_request(request, settings):
             continue
         
         if normalized in languages:
+            if debug:
+                log.info('Detect from HTTP Header=%s, lang=%s' %
+                         ('HTTP_ACCEPT_LANGUAGE', lang))
             return normalized
 
     #return default lanaguage